cambio en los reportes de inconsistencias

+++ Contasis/Clase/ruc.cs
@@ -95,26 +95,43 @@ namespace Contasis.Clase
             DataTable Tabla = new DataTable();
             SqlConnection cone = new SqlConnection();
 
-            try
+            string query0 = "select cg_empresa.NOMEMPRESA as Empresa From cg_empemisor inner join cg_empresa on cg_empemisor.ccodrucemisor=CG_EMPRESA.ccodrucemisor "+
+                            " where  cg_empemisor.ccodrucemisor='" + Objet.ruc + "'";
+            cone = ConexionSql.Instancial().establecerconexion();
+            SqlCommand commando = new SqlCommand(query0, cone);
+            cone.Open();
+            DataTable dt = new DataTable();
+            SqlDataAdapter data = new SqlDataAdapter(commando);
+            data.Fill(dt);
+            if (dt.Rows.Count > 0)
             {
-                string cadena1 = "Delete from cg_empemisor  where ccodrucemisor='" + Objet.ruc + "'";
-                cone = ConexionSql.Instancial().establecerconexion();
-                SqlCommand commando = new SqlCommand(cadena1, cone);
-                cone.Open();
-                cadena = commando.ExecuteNonQuery() == 1 ? "Eliminar" : "No se pudo eliminar";
+                MessageBox.Show("No puede eliminar este ruc esta activo con esta empresa : "+dt.Rows[0]["Empresa"].ToString()  ,"Contasis Corp.", MessageBoxButtons.OK, MessageBoxIcon.Stop);
             }
-            catch (Exception ex1)
+            else
             {
-                MessageBox.Show(ex1.ToString());
-                cadena = ex1.Message;
-            }
-            finally
-            {
-                if (cone.State == ConnectionState.Open)
-                {
-                    cone.Close();
-                }
 
+                try
+                {
+                    string cadena1 = "Delete from cg_empemisor  where ccodrucemisor='" + Objet.ruc + "'";
+                    cone = ConexionSql.Instancial().establecerconexion();
+                    SqlCommand commando1 = new SqlCommand(cadena1, cone);
+                    cone.Open();
+                    cadena = commando1.ExecuteNonQuery() == 1 ? "Eliminar" : "No se pudo eliminar";
+                }
+                catch (Exception ex1)
+                {
+                    MessageBox.Show(ex1.ToString());
+                    cadena = ex1.Message;
+                }
+                finally
+                {
+                    if (cone.State == ConnectionState.Open)
+                    {
+                        cone.Close();
+                    }
+
+                }
+                
             }
             return cadena;
         }
